requirements.txt
yfinance~=0.1.72
matplotlib~=3.5.2
discord~=1.7.3
python-dateutil~=2.8.2
python-dotenv~=0.20.0
requests~=2.28.0
fuzzywuzzy~=0.18.0